+++ buession-logging-springboot/src/main/java/com/buession/logging/springboot/autoconfigure/console/package-info.java
@@ -19,11 +19,11 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2023 Buession.com Inc.														       |
+ * | Copyright @ 2013-2024 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 /**
  * @author Yong.Teng
  * @since 0.0.1
  */
-package com.buession.logging.jdbc.support;
+package com.buession.logging.springboot.autoconfigure;