+++ buession-logging-support-console/src/main/java/com/buession/logging/console/formatter/DefaultConsoleLogDataFormatter.java
@@ -22,8 +22,43 @@
  * | Copyright @ 2013-2024 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
+package com.buession.logging.console.formatter;
+
+import com.buession.core.utils.StringUtils;
+import com.buession.logging.core.LogData;
+
 /**
+ * 默认控制台日志格式化
+ *
  * @author Yong.Teng
- * @since 0.0.1
+ * @since 0.0.4
  */
-package com.buession.logging.jdbc;
+public class DefaultConsoleLogDataFormatter implements ConsoleLogDataFormatter<String> {
+
+	@Override
+	public String format(final LogData logData) {
+		return logData == null ? null : logData.toString();
+	}
+
+	@Override
+	public String format(final String template, final LogData logData) {
+		if(logData == null){
+			return null;
+		}
+
+		String message = template;
+
+		message = StringUtils.replace(message, "${principal}", logData.getPrincipal().toString());
+		message = StringUtils.replace(message, "${time}", logData.getDateTime().toString());
+		message = StringUtils.replace(message, "${clientIp}", logData.getClientIp());
+		message = StringUtils.replace(message, "${User-Agent}", logData.getUserAgent());
+		message = StringUtils.replace(message, "${os_name}", logData.getOperatingSystem().getName());
+		message = StringUtils.replace(message, "${os_version}", logData.getOperatingSystem().getVersion());
+		message = StringUtils.replace(message, "${device_type}", logData.getDeviceType().getName());
+		message = StringUtils.replace(message, "${browser_name}", logData.getBrowser().getName());
+		message = StringUtils.replace(message, "${browser_version}", logData.getBrowser().getVersion());
+
+		return message;
+	}
+
+}