+++ buession-logging-support-console/src/main/java/com/buession/logging/console/package-info.java
@@ -26,4 +26,4 @@
  * @author Yong.Teng
  * @since 0.0.1
  */
-package com.buession.logging.jdbc;
+package com.buession.logging.console;